Add a service to create announcements
This service creates announcements for the landing page and stores them
in the database.

At the moment only the landing page is allowed to have announcements, so
this has been hardcoded in the service, rather than expecting the
service to be initialised with a CoronavirusPage object.

All of the current "live" announcements have a published date. Also
we're forcing users to add a "published_at" date when they create an
announcement, so the test date has been updated to reflect this.