Add service to discard draft announcements
Draft coronavirus page content is currently discarded using methods in
the Content Builder class.
The Content Builder class is therefore being used for two purposes:

1. Populate the Coronavirus pages with data from the YAML file.
2. Discard coronavirus page data and replace it with published version
in publishing-api

There might be other things it's doing, but these two at least can be
split up. Moving the "discarding" functionality to another class will
make both classes easier to update and maintain.

I've started by adding announcements to the new class so that they can
be used to establish the pattern for the class and the tests, before
unpicking ContentBuilder in later commits.