Call DraftDiscarder from ContentBuilder
Doing this gives confidence that we aren't losing any functionality as we move the discard
draft code from ContentBuilder to DraftDiscarder as the ContentBuilder tests should all still
pass. That is why the test being added here is almost a duplicate of the one previously added for
in the DraftDiscarder tests.
The tests will be removed from Content Builder when we are ready to call DraftDiscarder from
the controller directly.