Move discarding of sub_sections to DraftDiscarder
Only the code to do the discarding is being moved, so there is no change
to how SectionsPresenter is called.

Also replicates the ContentBuilder tests for discarding sub_sections so
we can be confident that no functionality has been lost.

This change also incorporates a bug fix where sub_sections are being
removed when the user clicks on "discard draft", but they are not being
replaced by the entries in publishing-api. For some reason this
behaviour was fixed by assigning
[sub_sections to a variable](https://github.com/alphagov/collections-publisher/pull/1178/commits/4f6f0566852b84c54580d49971bb6b4aff7deeb6),
so that change has also been replicated here.