Call publishing-api directly from DraftDiscarder
Now that all of the discarding of drafts is happening in DraftDiscarder
we should move the getting off the "live" content there as well rather
than passing it in.

The draft discarder tests have had to be updated to make sure that the
content_id of the coronavirus page matches the one in the publishing-api
payload otherwise the wrong data will be stubbed.