Get the "live" version of the content item
By default, `get_content` retrieves the latest version regardless of publication state. Though we will
be sending a discard-draft request to publishing-api, this request may not have completed by the
time we are ready to revert the coronavirus pate, so we should make sure that we are over-writing it with
the correct version of the content.