Add form validation and error mesages
Add nil check into href_format method to pass tests.'should validate_presence_of(:coronavirus_page)' in the announcement_spec.rb sets all attributes of announcement to nil and will fail if we don't check for a nil value.

More information on this can be found here
https://stackoverflow.com/questions/16175559/rails-rspec-shoulda-validate-presence-of-failed-with-setter-called-twice/16178678#16178678

We need a valid year check to prevent users from entering an invalid year, in this case we have decided that users should not be able to input a date in the future or too far in the past. The error handling in the AnnouncementHelper#format_published_at does not prevent this in the same way that it would for days and months

A side effect of the error handling in the announcement helper is that it rescues the error with nil. This made it difficult to check the presence of published_at because we couldn't differentiate between a completely empty date and a date that was rescued. We decided to just use the published_at_format validation to display for both cases.

We have adjusted the ErrorItemsHelper to display more user friendly field names rather than using the database field names. There is currently not a test for this helper, we plan to add this in a subsequent piece of work.

We also changed the default href value in the announcement factory to reflect the new requirements of the href validation i.e it must begin with a /.

Co-authored-by: Rosa Fox <rosa.fox@digital.cabinet-office.gov.uk>