Wrap discard tasks in a transaction
At the moment the "discard draft" action is failing. All of the subsections are being
removed, but they are not being replaced by the entries in publishing-api.

Error:
```
{"exception_class":"ActiveRecord::RecordNotSaved","exception_message":"Failed to replace sub_sections because one or more of the new records could not be saved.","stacktrace":["app/services/coronavirus_pages/model_builder.rb:40:in `store_live_subsections'","app/services/coronavirus_pages/model_builder.rb:23:in `discard_changes'","app/controllers/coronavirus_pages_controller.rb:64:in `discard_model_changes'","app/controllers/coronavirus_pages_controller.rb:39:in `discard'"]}
```

The hope is that wrapping the task in a transaction will stop all of the data from being removed
whilst we investigate what is happening.