Explicitly set sub_section attributes before save
Explicitly set sub_section attributes before trying to save multiple
sub_sections to a CoronavirusPage.

This is a pattern that is also used when discarding draft
[Step by Steps](https://github.com/alphagov/collections-publisher/blob/master/app/services/step_by_step_page_reverter.rb#L48-L61)

I have tried creating the sub_sections without storing them in the `new_sub_sections` variable first,
but that still bombs with the same error:

```
{"exception_class":"ActiveRecord::RecordNotSaved","exception_message":"Failed to replace sub_sections because one or more of the new records could not be saved.","stacktrace":["app/services/coronavirus_pages/model_builder.rb:40:in `store_live_subsections'","app/services/coronavirus_pages/model_builder.rb:23:in `discard_changes'","app/controllers/coronavirus_pages_controller.rb:64:in `discard_model_changes'","app/controllers/coronavirus_pages_controller.rb:39:in `discard'"]}
```

This is the only way I've found of fixing it.