Wrap discard tasks in a transaction
This removes the risk of some of the data (for example sub_sections) being removed,
but not recreated if saving to the database fails. So avoids the situation of changes
to announcements being discard successfully and replaced with the version on publishing-api,
but the sub_sections being deleted and not replaced.