Include step content in the body field
We want step content to be indexed by search so we can find these pages more
easily. Search-api can understand html, and steps are written in markdown.

We therefore want render all the step content and titles as html, not for
public consumption, but so that search can index the page content.

I've tried to present the output fairly verbosely with plenty of space between
sections. Fully fledged step by step pages are quite chunky, and should we
experience problems it'd be nice to be able to view this bit in a debugger
without it just being a mass of html.