Add missing tests for SpecialRoutePublisher
This was manually tested in https://github.com/alphagov/collections-publisher/pull/534,
but is missing automated tests. That made it the least covered
.rb file in this repository, at a coverage rate of 54.55%.

This commit adds the missing tests, which brings the overall
coverage rate of the repository up from 94.13% to 94.32%. We need
to aim for 95% before we can enable CD on this repo.