Continue to get announcements from YAML
Until we are ready to go live with the announcements section and import the existing
announcements from the Github to the database, we need to make sure that we are still
sending announcements to publishing-api correctly.

When we [updated the DraftUpdater](https://github.com/alphagov/collections-publisher/commit/52e8fc1e31ca5b2a3140be54ad10baf7f80f9c19#) to get the announcements from the
database we inadvertently removed the announcements from the publishing-api.
This is because the assignment `data["announcements"] = announcements_data` overwrites
the announcements pulled in from the github data, but at the moment there is nothing to replace it
with.

To get the tests to work, we need to update the fixture to not return any announcements. This is so that
it matches the assertion in step `then_the_reordered_subsections_are_sent_to_publishing_api`