Add feature flag for backend code
The announcements UI was being hidden behind a user permission
"Unreleased feature" to stop them seeing and interacting with
announcements, however there was nothing to stop the backend code from
running.

This caused problems recently where the announcements section was
removed [1] in live because the code to populate publishing-api was
looking for the announcements in the wrong place. We added a flag to
check if there are any existing announcements to get around the problem
but since found that this had a knock-on effect when drafts are
discarded.

When the draft is discarded, announcements are retrieved from
publishing-api and populated in the database. As the local database had
announcements, any changes made to the announcements section via Github
weren't picked up because the flag we added evaluated to true.

To stop problems like occurring in the upcoming timeline feature, config
has been added that sets the conditions for when the backend code
associated with an unreleased feature should be run.

The config class relies on the presence of an environment variable. The
bonus of this is that we can enable unreleased_features differently in
different environments.

For example, if we set ENV["UNRELEASED_FEATURES"] in integration only,
that means we'll be able to test the feature in Integration, but it
would be turned off by default in staging and production.

So rather than doing something like:

```
data["announcements"] = announcements_data if coronavirus_page.announcements.any?
```

We can instead do:

```
data["announcements"] = announcements_data if Rails.application.config.unreleased_features?
```

[1]: https://github.com/alphagov/collections-publisher/pull/1191