Simplify special route publisher job
By inverting the merge we can choose what type of route to publish in
the payload of each route if we wish. We should definitely default to
`exact` because `prefix` routes should generally be avoided unless there's
a really good reason for them.

I also don't think there's much need for us to be able to pass custom loggers
and publishing-api adapters into the publisher. It's just more stuff to
worry about.