Test ListItemsController's #create method
In an attempt to increase code coverage for this repository, I'm
tackling the least tested controller ('47.92% lines covered' at
time of writing).

The 'create' method was missing tests for 'flash[:danger]' and
the updateURL logic. In writing the tests, I discovered the JS
error response contained invalid JSON as the 'errors' JSON was
double encoded. I've now fixed this.