Add translated pages for /eubusiness
We'll need to wait for gds-api-adapters to be updated to 67.2.0 in order
to publish locales using SpecialRoutePublisher.

I've removed the public_updated_at entry as this is set by default downstream.

document_type is "special_route" by default, which is probably fine for
entirely hard-coded pages that are not referenced in search in any way.

I've left the content_ids of each of the translations the same, as this
indicates that they differ only by their locale, and should be linked together
as translations of each other should we wish to use this to help with our
navigation between translations.

I've added a task to publish a single page, as this seems useful. Unpublishing
them is more a moveable feast, so I haven't built anything for that yet.
I think the most likely outcome is that we want to redirect to the English version.