Test ListItemsController's #update method
The 'update' method was missing tests. It had the same issue as
the 'create' method, in that the returned JSON was invalid.

This is the only method that only supports JS (JSON) as a format,
so there are no tests for a HTML request format.

There's no 'subject' to this set of tests. Instead, there's a
'patch_list_item' method, because in the 'mark the tag as dirty'
test we need to check the value before and after the PATCH call.
We could do this by calling 'subject' in every test, but that
isn't as readable/self-documenting as calling 'patch_list_item'.