Feature spec for deleting announcements
- Requires JS due to confirmation modal.

- set_up_github_data has been moved within the feature step method definitions.
Previously it was called within coronavirus_page_spec which was
not fitting as it isn't a step that a user takes.

- It also lead to a clash in stubbing between the create and delete tests due
 to wanting to use given_there_is_coronavirus_page_with_announcements to set
up the delete tests. Because the set_up_github_data call happened before all the
create announcement feature step definitions were called, it ended up in a
circle of pain.

- Moving the set_up_github_data call to the methods in which the calls to publishing-api
 are triggered means that it doesn't effect other methods when it doesn't need to.

Co-authored-by: Leena Gupte <leena.gupte@digital.cabinet-office.gov.uk>