Test ListItemsController's #destroy method
Monitoring the calls to the list item and tag within
ListItemsController was a little hairy; I've had to write some
extensive stubs in places.

I've adopted the 'stub_const' approach for stubbing Tag, as
otherwise we'd risk polluting the global namespace and affecting
the outcome of other tests.

This final addition of tests brings us up to 95.2% coverage.
Previous figures should be taken with a pinch of salt, as this
branch has been rebased against master a couple of times in
development, so code coverage will fluctuate in response to other
peoples' changes.