Remove need for at least one live stream in db
Adds a guard so that the merging of live stream data doesn't fail if the
database doesn't contain a live stream.

The live stream is an optional component on the coronavirus landing
page, and there hasn't been a live stream in months.

However the code expected there to be at least one live stream in the
database.

Adding the guard means that we no longer have to create an unnecessary
live stream object to test things like announcements and sub_sections.