Add Timeline enty model and migration.
We need somewhere to store the timeline entries from the coronavirus_landing_page.yml within the publishing application.

It turns out that rails can handle pluralizing ‘entry’ to ‘entries’ so we shouldn’t have to specify the association between the model and the table

Has basic validation to check presence of fields. These fields don't currently match the fields in publishing-api, but that doesn't matter right now. We can do the mapping later when it's time to update the payload to publishing-api.

Currently there is no check for the presence of a 'position' value and
we have no way of currently assigning a value. This will need to be
adressed in subsequent work as the table will not accept a null value
for 'position'. I have not set a default value for position as the
intention is that this value will be incremented for each new value.