Hide bullet button in markdown editor by default
More options are going to be added to the markdown editor in later
commits.

Rather than having to set `hide_...` to true to turn them off and having
a lot of `unless` statements in the view, it will be easier to follow
and update if these statements are truthy instead.

All of the new options will be turned off by default. The calling view
will need to "turn on" the options/buttons it wants to give users access
to.

This change in logic means that we no longer have to hide the bullets
button on the sub section page, but we do need to explicitly turn them
on for step by steps.