Create form to enter timeline entry details
`form_with` is used there as it is replacing `form_tag` and `form_for`
which have been  deprecated.
One of the quirks of `form_with` though is that by default `remote` is
set to true [1][2]. If you don't override this setting with
`local: true` the form won't submit.

Rather than setting `local: true` in the call to `form_with`, this
config is being set globally for all forms that use `form_with`. This
will prevent others from being caught out by the form not submitting,
and it means we only need to remove one line of config after the
eventual upgrade to Rails 6.1 where the default is being set back to
`local` [3].

See:
[1]: https://edgeguides.rubyonrails.org/form_helpers.html#how-do-forms-with-patch-put-or-delete-methods-work-questionmark
[2]: https://stackoverflow.com/questions/51892516/rails-form-with-method-get-cannot-update-a-view/51893521
[3]: https://github.com/rails/rails/pull/40946