Include timeline entries in Coronavirus page content
This uses data from the timeline_entries association of the
coronavirus_page to populate the data sent to the Publishing API. This
is put behind the unreleased_features config option as this should only
be done once the DB is populated in production.

A minor pain point with the modelling of TimelineEntry is the mismatch
between Publishing API using paragraph and the model using content,
without this things would be simpler - but I've not tried to address it
here.

In order to test the feature without unreleased_features enabled I've
had to set deep_dup when `data` is defined in the spec. This means that
any changes to `data` do not also alter the `github_content` and allow
it to be re-used in a test.

Adding the timeline_data test seemed a little strange as that seems to
be a private method but this class seems to lack those - I just followed
existing convention here despite it feeling a bit wrong.