Change TimelineEntries#content to be text instead of string
A string type becomes a VARCHAR(255) in MySQL which is shorter than what
is normally used for an update (using the current one causes an
exception).

Setting this to text sets the limit to 64KB which should meet all our
expectations for how a content designer can reasonably use this field.