Validate length of timeline entry heading
Since 255 characters isn't all that long before you get a DB exception
it seems prudent to have a validation here so the user experience is OK
if/when someone enters a long heading.

An alternative to this could be to set the column to be a bigger one and
not validate a limit, however it seems that it'd be rather poor for the
page visually to use > 255 characters for a heading.