Allow null positions in timeline entry
Position will be used to determine the order in which timeline entries are displayed to users.
However that ordering is happening in a later story. At the moment we just need to allow users
to enter a heading and some content and be able to save to the database, so position is being
changed to allow nulls.