Pass list of "controls" to markdown editor
Pass a list of controls to turn on to the markdown editor rather than
individual booleans.

The timeline editor is going to allow users to see the full range of
toolbar buttons, e.g. headings, blockquotes etc.

Rather than passing this in as an individual list of booleans, which
could get quite long and unwieldy, a list of allowed "controls" will be
passed in instead.

If no controls are passed in, just the "list" button will be displayed
as that is the only one outside of a conditional, however, to see more,
the controls will have to be explicitly passed in in the views.
e.g. `controls: [:headings, :blockquotes, :numbered_list]`