Restrict timeline entries to "Unreleased feature"
Implementing timeline entries is still a work in progress,
so we should restrict access to it to users with
"Unreleased feature" permissions.

I also added in the 'topic page?' check to make sure that
the timeline entries section was visible on the landing page
but not on any of the sub section pages.