Add position value when timeline entry is created
By default, the new timeline entry will be added to the top of the list,
i.e. position 1, so that they appear in the correct order on the summary
page and landing page preview, so that content designers can be sure
that the order of the entries they see in the publishing tool matches
the order of the timeline entries shown to the user.

For example, a content designer might add an entry for "2 January", and
later add an entry for "24 - 26 January". By default these entries in
will be displayed in this order:

"24 -26 January"
"2 January"

The "position" has been removed from the factory to match what occurs
when a new timeline entry is created by the user. i.e. it is created
with a position of `nil`.
The position is added by the `after_create` callback.
If we leave the default value in the factory, when `after_create` runs
the position will be set to 2, even though there's only one timeline
entry.

This solution isn't the most elegant, but it has the benefit of
unblocking the rest of the timeline entries work. The other solution
considered was to use the [acts_as_list][1] gem, but this was taking a
while to implement. So the decision was made to finish the rest of the
timeline feature, and then look at improving reordering for both
announcements and timeline entries.

[1]: https://github.com/brendon/acts_as_list