Update timeline positions when one is deleted
The position numbers should be sequential. When a timeline entry is
removed from the beginning or the middle of the list the timeline
entries "below" it should be moved up a position.

destroy_all changes to delete_all in the draft discarder so that the
after_destroy hook in the model isn't called. This is because the
positions are set when the timeline entries are created in the draft
discarder.