Add timeline entries section
Add a partial for the new timeline entries section. This follows the
pattern established in the announcements and subsection partials.

The links to the "Reorder" page, "Add timeline entry" page, edit page and delete button don't go
anywhere yet. They will need to be updated when these pages/functions are built in subsequent work.