Refactor landing page summary list css
It is unnecessary to have separate css classes for the individual sections as much of the styling for them is the same. Here we have consolidated the styles into a summary-list-divider class which means we don't have to have classes for accordion, announcement and timeline entries that only differ slightly in their styles. It reduces the amount of repetition significantly.