Add a feature test for timeline entries
An assumption has been made that the work in PR #1253 [1] will be added
in some form, and that newer timeline lines will be set with position 1.
That's why the positions of the test timeline entries are being
explicitly set in the tests, so that there is no ambiguity about the
starting and ending order of the timeline entries.

[1]: https://github.com/alphagov/collections-publisher/pull/1253