DRY-up test helper method
A `setup_github_data` method was being created in multiple test files,
but they all bascically did the same thing. This use of this method has
been replaced by the `stub_coronavirus_landing_page_content` in a new
test helper module.

To be able to use this method, `CoronavirusHelpers` needs to be included
in the tests. This exposed a problem with the Coronavirus Feature Steps
methods as they were not in a module, and therefore could not include
other helpers. To get around this the Coronavirus Feature Steps methods
were added to a module. This means that the
`CoronavirusFeatureSteps` helper now needs to be included every where
one of it's methods' is used. This wasn't problem before because methods
outside of a module are included in every Ruby object.