Only set featured link descriptions for relative links
We want to allow content designers to feature links that are to external
website (such as NHS). This was previously blocked due to needing to
pull a description from the Publishing API which required a relative
path. This adds a guard clause that doesn't pull in description for an
absolute link

It also does some refactoring to introduce the extra test so that they
are more succinct.

Co-authored-by: Alex Newton <alex.newton@digital.cabinet-office.gov.uk>
Co-authored-by: Rosa Fox <rosa.fox@digital.cabinet-office.gov.uk>