Get Coronavirus page from DB for sub-sections
The reorder pages for sub-sections was using ModelBuilder to get the details of the
coronavirus page. This is unnecessary as the coronavirus page is in the DB and we can
get the details from there.

Also removes the redirect to the landing page, if the reorder page doesn't exist. Firstly no
tests failed when this action was removed. Also, this is only necessary if we anticipate users
hacking the url and going straight to the reorder page, which is very unlikely.