Wrap destroy sub-section action in a transaction
Wrap the destroy action in a transaction so that it can be rolled back
if the call to publishing-api fails. This has the benefit of preserving
the record, including the current position of the sub-section.

The tests have been tied a little. A `let!` is being used to create the
sub-section so that it is perpetuated in the tests that need it.

The test also checks that the sub_sections haven't changed, i.e.
been deleted and recreated. To do this we need to change them
to an array to check the contents of each subsection otherwise the
following error is thrown:

```
Failure/Error: expect { subject }.to_not change { coronavirus_page.reload.sub_sections }
       expected `coronavirus_page.reload.sub_sections` not to have changed, but did change from #<ActiveRecord::Associations::CollectionProxy [#<SubSection id: 2, title: "Perspiciatis aut ratione v...ted_at: "2021-02-01 11:48:47", updated_at: "2021-02-01 11:48:47", position: 1, featured_link: nil>]> to #<ActiveRecord::Associations::CollectionProxy [#<SubSection id: 2, title: "Perspiciatis aut ratione v...ted_at: "2021-02-01 11:48:47", updated_at: "2021-02-01 11:48:47", position: 1, featured_link: nil>]>
```

This is because Rspec change[1] also checks the hash[2] method which
differ for collectionproxy objects but not arrays.

[1]: https://github.com/rspec/rspec-expectations/blob/bd10f0cf3970932781efcd09b5e427877d16a6c2/lib/rspec/matchers/built_in/change.rb#L393
[2]: https://ruby-doc.org/core-2.7.1/Object.html#method-i-hash