Put coronavirus models in a Coronavirus namespace
This applies a namespace of Coronavirus that can be used to
differentiate the various code that is related to the content management
of coronavirus.

This does leave the CoronavirusPage model named rather awkwardly since
it is now Coronavirus::CoronavirusPage. I plan to rename this to
Coronavirus::Page in a later commit. I also didn't touch the LiveStream
model since that concept of code seems to have a few oddities that are
best dealt with in isolation.

I wasn't sure of the best name to give the namespace, CoronavirusPages
seemed a good option but it conflicted with the model and controller of
the same name. A suggestion in the trello card was CoronavirusPublisher
but I felt that sounded more like a product than a namespace. I arrived
at just plain Coronavirus since that already matched the path used for
routes (/coronavirus).