Switch CoronavirusPages namespace to Coronavirus::Pages
This change has been made so that these classes could be within the
Coronavirus namespace. These classes didn't seem to make much sense
without some relationship to page so I decided to create an additional
namespace of Pages with Coronavirus so they could be moved with minimal
changes.

There is probably scope to rethink the purpose of some of these classes
as they seem to venture outside of what would be considered to be a
service (Configuration) and there is a Presenter class here that seems
to have cross over with the presenters directory of classes.