Change live stream routing to be under /coronavirus
This puts the landing page route under a coronavirus namespace to
reflect that it is part of the coronavirus code. This changes this route
to be prefixed with /coronavirus and expects the controller to be under
a Coronavirus namespace (which I'll do in the next commit).

This then changes the root route to redirect to /coronavirus/live_stream
for users with the "Livestream editor" permission rather than serve the
page directly. This is to be consistent with the /step-by-step-pages
route and avoids this route being available in two places.