Rename most usages of coronavirus_page in app
Now that this model is in a namespace of Coronavirus::Page it seems
unusual to refer to this in variables as coronavirus_page whereas other
models such as Coronavirus::Announcement or Coronavirus::SubSection are
referred to as announcement or sub_section. This systematically renames
the places where the use of `coronavirus_page` can be renamed to `page`
because it is within a Coronavirus context.

It leaves behind the feature tests as these are not in a coronavirus
context and would be confusing as just `page`.