Move coronavirus_pages routes inside coronavirus namespace
The resource name :coronavirus_pages needed changing to reflect that the
resource is now named :pages. To do this I've moved this within the
coronavirus namespace and set the path of the pages resource to be an
empty string to maintain the past URL.

This move meant that the route parameter was now page_slug rather than
coronavirus_page_slug which meant a bit of churn through the code to
edit all the references.

As part of this I've switched the publish route to have an `on: :member`
condition so that this route has a `slug` parameter like other routes on
that controller, this allows removing the slug conditional in the
controller. It also means the route is now named
`publish_coronavirus_page` rather than `coronavirus_pages_publish` which
feels more consistent.