Add coronavirus prefix to coronavirus database tables
This makes these models consistent with `coronavirus_pages` in having a
`coronavirus_` prefix in the table name.

This migration must be applied at a quiet time for the application as
the deployment of it risks errors from when the migration has run until
a new instance of the app is running.

It's not ideal to deal with this risk of errors but the alternative is
far more painful - which would be creating new tables that are
backfilled and kept in sync until the tables are swapped. Given that
this application frequently goes many hours without a user request it
seems reasonable to take this risk.