Restructure the content on the index page
Currently the content is is various places in the locale file and it's not obvious where to find
the content for a page. The current plan is to structure the content so that it follows the flow
of the pages from the index page. That means removing the confusing "pages" heading. Whilst
that is the name of the controller, it doesn't really mean much as far as the page flow is concerned.