Restructure instructions for updating from github
This content is all over the place. It looks like the content is being used in shared
partials, but those partials are only being used by the "prepare" view, which for the
user is the page to "edit something else". The "shared" folder will be renamed in the
next commit to make it clear that they are only being used by one page.

The content for the instructions is being pulled together in one place to make it easier to
find and update, similar to the live stream instructions.