Fix subsections bug
In the previous configuration we had accidentally allowed the user to
save changes to a sub-section even if there were errors with it. This
wraps the sub-section changes in a transaction whilst we call on the
draft updater, so they can be rolled back in the event that the draft
updater finds errors.