Raise exception for missing livestream translation
We already have `config.action_view.raise_on_missing_translations = true` set
so an error will be raised if a view tries to call a missing translation.
However, to get the same effect with content called from a controller, the method needs
to be called with a `!`. i.e. `I18n.t!`.
Rather than call translations without a bang in the views, but with one in the controllers,
we can instead use the helper method from the view.