Create new github_pages controller view and routes
When the functionality was first added to collections-publisher to allow users to
publish changes to the coronavirus pages, it made sense that the action to get the
changes ready for publish was called "prepare".

It made sense because all of the content was updated in github and pulled across to
collections publisher to be published, but wasn’t actually being stored.

Now, though with content being stored and updated in collections-publisher directly,
the name "prepare" is confusing.

When the "prepare" action was in the PagesController it wasn't clear that it was the only
thing that uses the "update" route. Now that that the page is being split it should
be very clear what this page is doing.

The name "github changes" was chosen to reflect that this route is only concerned
about getting and updating the content from Github. If the user wants to update
content in the timeline etc, they need to go elsewhere.

The "prepare" code will be removed from the PagesController in the next commit.