Fix subsection bug for the create method.
This fixes a bug whereby it was possible to save a new subsection with
invalid content. It's also brought to light a lack of error handling
that we have not only for the create method but for the update as well
which will be updated in a subsequent commit.

We've been ignoring ActiveRecord validation errors and relying on the
validation that we have in the presenters to show users their errors.
We're checking the validity of sub_sections before we make any attempt
to save them. This means that we're not at any risk of saving invalid
subsections and that we are using errors as created by the model layer.