Remove "prepare" action from PagesController
Now that the actions to update content from Github has it's own
controller etc, we can remove the code lingering in the PagesController.

This change only removes the old "prepare" code. It doesn't try to clean
up the PagesController further as that is bigger job and outside the
scope of this particular change. The same goes for cleaning up the test
files.