Throw not found error for coronavirus page
If the user hacked the url and entered a slug that doesn't exist, they were being redirected to
the index page.

Now that we are populating links to the summary pages from the database, this feels like odd
behaviour. Removing this redirect means that we can now just use `find_by!` to throw a 404.