Refactor subsection tests and set error status codes
This organises the tests for create and update actions to be based
around the 3 outcomes they can have as contexts - this has been done to
make the testing easier to follow and be closer coupled to the code
being tested.

In doing this I noticed we weren't returning a 422 status code when
there was invalid input so I've included that. I've also set the status
code to 500 for when the draft updater fails to indicate that we expect
that this is caused by a service error (this glosses over how it can
be caused by invalid user input as hopefully we'll fix this soon).

Note: the CGI.escapeHTML is needed because Rails HTML encodes the quote.