Add framework for ContentGroup model
As it stands we do the validation of our accordian content within the sub section json presenter.
This would ideally be done within a database model because it's presenting error messages to
the user. This is complicated somewhat by the fact that we are not offering a specific way in
which to enter content in the format that we want it (headings and links only) and are instead
breaking the input content into individual lines before ensuring that it matches a link or
heading regex. Moving all of this logic into the SubSection model would heavily bloat it and
give it functionality it shouldn't really have. For this reason we have instead decided to
create a new model that can validate individual headings and the list of links below them. This
means that we can move that validation slowly over to the new ContentGroup model, the framework
for which is in this commit.