Create content groups from the sub_section model
This makes it so that the SubSection model has responsibility for creating the new
ContentGroups. As it stands this is working alongside the existing implementation
in the Json Presenter but once the logic in the SubSections model is more complete
and working for update and destroy as well as Create we can look at moving the
things that users actually see over to the new way of validation.