Raise an error if subsection has invalid markdown
Changes SubSectionJsonPresenter to raise an error if the markdown is invalid.
The benefit of doing this is that we will eventually be able to surface this error
in the Subsection controller and rollback the creation and update of subsections
based on this error rather than having to raise an `ActiveRecord::Rollback` as we
are currently doing.