Raise error if page contains invalid content
This also takes the markdown invalid error from SubSectionJsonPresenter
and re-raises it as the more generic InvalidContentError. This is so
that callers of `ContentBuilder` don't need to worry about exceptions
raised in subclasses.

This error and the code that triggers it should be removed when the
validation is moved to the model.