Catch discarding errors in PagesController
Rather than checking a list of errors, we can catch the exceptions
raised by the discard method instead.

This is to keep things consistent with upcoming changes to the `send`
method which will also start raising exceptions, which will allow
changes made in the controllers to be automatically rolled back once we
start wrapping database calls in transactions.