Raise an error if updating the draft fails
The `send` method was the last method using the error_handler so that
can be removed too.
We don't have Sentry alerts configured and we're already relying on
people to tell us when they need dev support, so there doesn't seem to
be much point in logging these errors in Sentry.