Catch update errors in timeline entries controller
Update the timeline entries controller to catch errors from DraftUpdater
in the update method.
It's not much extra code to also wrap the actions in a transaction which
can be rollbacked if DraftUpdater fails.

The update action is displaying the error from DraftUpdater on fail
because there currently isn't a fail message for these actions, and we
should show the user something to let them know there has been a
failure.

The call to `save!` runs the validations, but if they fail, an exception
is thrown and displayed to the user.
To void this, the validations are being run before the transaction so
the user can be shown a nice friendly error.