Add structured content model and tests
This adds the framework for structured content. This validates the content at the point that it is passed in and will go on to replace the validation that is currently done by the json presenter.

It contains some simple class methods to validate content before a sub section is created. It also contains the regexs that will be reused by the json presenter down the line. The parse method is responsible for checking the format of lines passed and ultimately returns a list of links and headings in the same order that it's passed them.

Co-authored-by: Alex Newton <alex.newton@digital.cabinet-office.gov.uk>
Co-authored-by: Kevin Dew <kevin.dew@digital.cabinet-office.gov.uk>