Create structured content when creating a sub section
This makes it so that structured content is created when a sub section is. It also adds the validation into the sub section model to ensure that the content is valid according to the Structured content. This in turn means that it is impossible for invalid content to be saved to the database.

It also makes it so that when `content` is called on the `SubSection` model the associated `StructuredContent` is updated. This has been achieved by creating a new `content` method on the `SubSection`. It also ensures that that same population is done after the initialisation of the model. This is to make sure that `StructuredContent` always exists and is usable by the Json Presenter later on.

Co-authored-by: Alex Newton <alex.newton@digital.cabinet-office.gov.uk>
Co-authored-by: Kevin Dew <kevin.dew@digital.cabinet-office.gov.uk>